requirements.txt
django
django-formtools
django-filter
django-allauth[mfa,socialaccount]
django-tenants
django-htmx
python-dotenv
Pillow
psycopg2-binary
django-countries
django-phonenumber-field[phonenumbers]
babel
django-jazzmin
django-colorfield
django-debug-toolbar
django-extensions
django-widget-tweaks
django-compressor
django-tailwind
gunicorn
dj-database-url
whitenoise
whitenoise[brotli]
celery
requests